idle-assist.sh: Update spin_check()
spin_check() checked if the log is valid by checking the length of the log, but
it didn't check the validity of the grep'ed variables, possibly led to another
segmentation fault error or any other errors.

Updated spin_check() to check the validity of the variable and add routine to
bail out when recheck is failed 10 times in a row.