fix: Fix Segmentation fault by exceeding stack limit
Currently, main() function works as recursive function, making stacks over bash
stack size limit (8192). Correct fix for this segmentation fault error would be
rewrite the main code that does not rely on recursive function, but it will take
some time to do that. So, until then, just setting the stack limit to unlimited
will do its job. The only problem of this temporary fix is that the script can
take unlimited space in RAM, but that would not be a serious problem since this
script is not that complicated.

Commit 27b993681055a9d74bbce684022d5e93f00e5606 was not a correct fix for this.